refactor: rename PR files to camelCase
Rename PascalCase filenames touched in this PR to match the
project's camelCase file naming convention and update imports.
