+++ apps/webapp/src/script/components/Meeting/MeetingList/MeetingList.tsx
@@ -27,8 +27,8 @@ import {Loading} from '@wireapp/react-ui-kit';
 import {emptyListContainerStyles} from 'Components/Meeting/EmptyMeetingList/EmptyListStyles';
 import {EmptyMeetingList} from 'Components/Meeting/EmptyMeetingList/EmptyMeetingList';
 import {meetingListContainerStyles} from 'Components/Meeting/MeetingList/MeetingList.styles';
-import {MeetingListItemGroup} from 'Components/Meeting/MeetingList/MeetingListItemGroup/MeetingListItemGroup';
-import {TodayAndOngoingSection} from 'Components/Meeting/MeetingList/TodayAndOngoingSection/TodayAndOngoingSection';
+import {MeetingListItemGroup} from 'Components/Meeting/MeetingList/MeetingListItemGroup/meetingListItemGroup';
+import {TodayAndOngoingSection} from 'Components/Meeting/MeetingList/TodayAndOngoingSection/todayAndOngoingSection';
 import {partitionMeetingsByDay} from 'Components/Meeting/partitionMeetingsByDay';
 import type {ScheduleMeetingRecurrenceOption} from 'Components/Meeting/ScheduleMeetingModal/scheduleMeetingTypes';
 import {getTodayTomorrowLabels, groupByStartHour} from 'Components/Meeting/utils/MeetingDatesUtil';
