+++ apps/webapp/src/script/components/Meeting/MeetingList/MeetingListItemGroup/meetingListItemGroup.styles.ts
