+++ apps/webapp/src/script/components/Meeting/utils/meetingStatusUtil.ts
